Move subctl part to `cmd/subctl/cloud`
