+++ pkg/cloud/cloud.go
@@ -1,43 +1,26 @@
-/*
-SPDX-License-Identifier: Apache-2.0
-
-Copyright Contributors to the Submariner project.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 package cloud
 
-import (
-	"github.com/spf13/cobra"
-	"github.com/submariner-io/submariner-operator/internal/restconfig"
-	"github.com/submariner-io/submariner-operator/pkg/subctl/cmd"
-	"github.com/submariner-io/submariner-operator/pkg/subctl/cmd/cloud/cleanup"
-	"github.com/submariner-io/submariner-operator/pkg/subctl/cmd/cloud/prepare"
-)
-
-var (
-	cloudCmd = &cobra.Command{
-		Use:   "cloud",
-		Short: "Cloud operations",
-		Long:  `This command contains cloud operations related to Submariner installation.`,
-	}
-	restConfigProducer = restconfig.NewProducer()
-)
-
-func init() {
-	cloudCmd.AddCommand(prepare.NewCommand(&restConfigProducer))
-	cloudCmd.AddCommand(cleanup.NewCommand(&restConfigProducer))
-	restConfigProducer.AddKubeContextFlag(cloudCmd)
-	cmd.AddToRootCommand(cloudCmd)
+type Options struct {
+	InfraID         string
+	Region          string
+	Profile         string
+	CredentialsFile string
+	OcpMetadataFile string
+	ProjectID       string
 }
+
+type Ports struct {
+	Natt         uint16
+	NatDiscovery uint16
+	Vxlan        uint16
+	Metrics      uint16
+}
+
+type Instances struct {
+	AWSGWType string
+	GCPGWType string
+	Gateways          int
+	DedicatedGateway  bool
+}
+
+const DefaultNumGateways = 1
