+++ pkg/cloud/prepare/gcp.go
@@ -25,26 +25,7 @@ import (
 	"github.com/submariner-io/submariner-operator/pkg/subctl/cmd/utils"
 )
 
-// NewCommand returns a new cobra.Command used to prepare a cloud infrastructure.
-func newGCPPrepareCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "gcp",
-		Short: "Prepare an OpenShift GCP cloud",
-		Long:  "This command prepares an OpenShift installer-provisioned infrastructure (IPI) on GCP cloud for Submariner installation.",
-		Run:   prepareGCP,
-	}
-
-	gcp.AddGCPFlags(cmd)
-	cmd.Flags().StringVar(&gcpGWInstanceType, "gateway-instance", "n1-standard-4", "Type of gateway instance machine")
-	cmd.Flags().IntVar(&gateways, "gateways", DefaultNumGateways,
-		"Number of gateways to deploy")
-	cmd.Flags().BoolVar(&dedicatedGateway, "dedicated-gateway", false,
-		"Whether a dedicated gateway node has to be deployed (default false)")
-
-	return cmd
-}
-
-func prepareGCP(cmd *cobra.Command, args []string) {
+func GCP(cmd *cobra.Command, args []string) {
 	gwPorts := []api.PortSpec{
 		{Port: nattPort, Protocol: "udp"},
 		{Port: natDiscoveryPort, Protocol: "udp"},
